Add go docs and update README.md
