+++ v2/context.go
@@ -2,6 +2,8 @@ package uwe
 
 import "context"
 
+// Context is a wrapper over the standard `context.Context`.
+// The main purpose of this is to extend in the future.
 type Context interface {
 	context.Context
 }
@@ -10,6 +12,7 @@ type ctx struct {
 	context.Context
 }
 
+// NewContext returns new context.
 func NewContext() Context {
 	return ctx{
 		Context: context.Background(),
