+++ v2/presets/api/server.go
@@ -11,7 +11,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-const ForceStopTimeout = 5 * time.Second
+const forceStopTimeout = 5 * time.Second
 
 // Config is a parameters for `http.Server`.
 type Config struct {
@@ -34,12 +34,17 @@ func (c *Config) TCPAddr() string {
 	return fmt.Sprintf("%s:%d", c.Host, c.Port)
 }
 
-// Server
+// Server is worker by default for starting a standard HTTP server.
+// Server requires configuration and filled `http.Handler`.
+// The HTTP server will work properly and will be correctly disconnected upon a signal from Supervisor (Chief).
+// Warning: this Server does not process SSL/TLS certificates on its own.
+// 		To start an HTTPS server, look for a specific worker.
 type Server struct {
 	config Config
 	router http.Handler
 }
 
+// NewServer returns a new instance of `Server` with the passed configuration and HTTP router.
 func NewServer(config Config, router http.Handler) *Server {
 	return &Server{
 		config: config,
@@ -47,14 +52,13 @@ func NewServer(config Config, router http.Handler) *Server {
 	}
 }
 
-func (s *Server) Init() error {
-	return nil
-}
+// Init is a method to satisfy `uwe.Worker` interface.
+func (s *Server) Init() error { return nil }
 
+// Run starts serving the passed `http.Handler` with HTTP server.
 func (s *Server) Run(ctx uwe.Context) error {
-	addr := fmt.Sprintf("%s:%d", s.config.Host, s.config.Port)
 	server := &http.Server{
-		Addr:    addr,
+		Addr:    s.config.TCPAddr(),
 		Handler: s.router,
 	}
 
@@ -68,7 +72,7 @@ func (s *Server) Run(ctx uwe.Context) error {
 
 	select {
 	case <-ctx.Done():
-		serverCtx, cancel := context.WithTimeout(context.Background(), ForceStopTimeout)
+		serverCtx, cancel := context.WithTimeout(context.Background(), forceStopTimeout)
 		defer cancel()
 
 		err := server.Shutdown(serverCtx)
