+++ v2/utils.go
@@ -2,11 +2,14 @@ package uwe
 
 import (
 	"errors"
+	"fmt"
+	"log"
 	"reflect"
 
 	"github.com/sirupsen/logrus"
 )
 
+// WorkerExistRule is a custom validation rule for the `ozzo-validation` package.
 type WorkerExistRule struct {
 	message          string
 	AvailableWorkers map[WorkerName]struct{}
@@ -36,7 +39,7 @@ func (r *WorkerExistRule) Error(message string) *WorkerExistRule {
 	}
 }
 
-// LogrusEventHandler returns default `EventHandler` that can be used for `Chief.SetEventHandler()`.
+// LogrusEventHandler returns default `EventHandler` that can be used for `Chief.SetEventHandler(...)`.
 func LogrusEventHandler(entry *logrus.Entry) EventHandler {
 	return func(event Event) {
 		var level logrus.Level
@@ -50,5 +53,22 @@ func LogrusEventHandler(entry *logrus.Entry) EventHandler {
 		}
 
 		entry.WithFields(event.Fields).Log(level, event.Message)
+	}
+}
+
+// STDLogEventHandler returns a callback that handles internal `Chief` events and logs its.
+func STDLogEventHandler() func(event Event) {
+	return func(event Event) {
+		var level string
+		switch event.Level {
+		case LvlFatal, LvlError:
+			level = "ERROR"
+		case LvlInfo:
+			level = "INFO"
+		default:
+			level = "WARN"
+		}
+
+		log.Println(fmt.Sprintf("%s: %s %+v", level, event.Message, event.Fields))
 	}
 }
