+++ v2/worker.go
@@ -3,9 +3,8 @@ package uwe
 import (
 	"context"
 
-	"github.com/pkg/errors"
-
 	"github.com/lancer-kit/sam"
+	"github.com/pkg/errors"
 )
 
 type WorkerName string
@@ -13,11 +12,14 @@ type WorkerName string
 // Worker is an interface for async workers
 // which launches and manages by the `Chief`.
 type Worker interface {
-	// Init initializes new instance of the `Worker` implementation
+	// Init initializes some state of the worker that required interaction with outer context,
+	// for example, initialize some connectors. In many cases this method is optional,
+	// so it can be implemented as empty: `func (*W) Init() error { return nil }`.
 	Init() error
-	// Run starts the `Worker` instance execution.
-	// Context should be used for listening to ctx.Done()
-	Run(Context) error
+
+	// Run  starts the `Worker` instance execution. The context will provide a signal
+	// when a worker must stop through the `ctx.Done()`.
+	Run(ctx Context) error
 }
 
 // workerRO worker runtime object, hold worker instance, state and communication chanel
@@ -41,10 +43,11 @@ const (
 // (*) -> [New] -> [Initialized] -> [Run] -> [Stopped]
 //          |             |           |
 //          |             |           ↓
-//          |_____________|------> [Failed]
+//          |-------------|------> [Failed]
 func newWorkerSM() (sam.StateMachine, error) {
 	sm := sam.NewStateMachine()
 	s := &sm
+
 	workerSM, err := s.
 		AddTransitions(WStateNew, WStateInitialized, WStateFailed).
 		AddTransitions(WStateInitialized, WStateRun, WStateFailed).
@@ -53,6 +56,7 @@ func newWorkerSM() (sam.StateMachine, error) {
 	if err != nil || workerSM == nil {
 		return sm, errors.Wrap(err, "worker state machine init failed: ")
 	}
+
 	if err = workerSM.SetState(WStateNew); err != nil {
 		return sm, errors.Wrap(err, "failed to set state new")
 	}
